requirement.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas
langchain_openai
python-multipart
jinja2
python-dotenv
langchain
langchain_core

